fix integer overflow in delta calculations.

+++ src/stats_thread.h
@@ -3,14 +3,14 @@
 
 struct sample {
 	struct timespec timestamp;
-	uint64_t 	rx_bytes;
-	uint64_t 	tx_bytes;
-	uint32_t 	rx_bytes_delta;
-	uint32_t 	tx_bytes_delta;
-	uint32_t 	rx_packets;
-	uint32_t 	rx_packets_delta;
-	uint32_t 	tx_packets;
-	uint32_t 	tx_packets_delta;
+	int64_t 	rx_bytes;
+	int64_t 	tx_bytes;
+	int64_t 	rx_bytes_delta;
+	int64_t 	tx_bytes_delta;
+	int64_t 	rx_packets;
+	int64_t 	rx_packets_delta;
+	int64_t 	tx_packets;
+	int64_t 	tx_packets_delta;
 };
 
 struct iface_stats {
